+++ xds/internal/xdsclient/xdsresource/route_config_resource_type.go
@@ -19,7 +19,6 @@ package xdsresource
 
 import (
 	"google.golang.org/grpc/internal/pretty"
-	"google.golang.org/grpc/xds/internal"
 	"google.golang.org/grpc/xds/internal/xdsclient/xdsresource/version"
 	"google.golang.org/protobuf/proto"
 	"google.golang.org/protobuf/types/known/anypb"
@@ -46,14 +45,6 @@ var (
 // Implements the Type interface.
 type routeConfigResourceType struct {
 	resourceTypeState
-}
-
-func init() {
-	if internal.ResourceTypeMapForTesting == nil {
-		internal.ResourceTypeMapForTesting = make(map[string]interface{})
-	}
-	internal.ResourceTypeMapForTesting[version.V3RouteConfigURL] = routeConfigType
-
 }
 
 // Decode deserializes and validates an xDS resource serialized inside the
